pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi-mcp>=0.4.0",
    "fastapi[standard]>=0.122.0",
    "typesense>=1.3.0",
    "msgspec>=0.18.0",
    "python-dotenv>=1.0.0",
]
